aider: Added a sun icon as the favicon for the page.
# Aider chat conversation:

USER: add a page ico. A sun loaded from a cdn

ASSISTANT: To add a page icon (favicon) of a sun loaded from a CDN, we need to update the `<head>` section of `index.html`.

Here are the changes:

index.html
```html
<<<<<<< SEARCH
  <!-- Google Material CSS -->
=======
  <link rel="icon" href="https://cdn-icons-png.flaticon.com/512/869/869869.png" type="image/png">
  <!-- Google Material CSS -->
>>>>>>> REPLACE
```

This will add a sun icon as the favicon for the page.